pyproject.toml
[project]
name = "dash-umap-image-explorer"
version = "0.1.0"
description = "Interactive UMAP-based image exploration with Dash and DMC"
requires-python = ">=3.9"
dependencies = [
    "dash[diskcache]>=2.18.0",
    "dash-mantine-components>=0.14.0",
    "dash-ag-grid>=31.0.0",
    "plotly>=5.24.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.5.0",
    "umap-learn>=0.5.7",
    "pynndescent>=0.5.14",
    "numba>=0.60.0",
    "pillow>=10.4.0",
    "flask-caching>=2.3.0",
    "diskcache>=5.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = []
